Move picker/date picker specifics into a module.
