+++ lib/formotion/row_type/multi_choice_row.rb
@@ -1,0 +1,25 @@
+module Formotion
+  module RowType
+    module MultiChoiceRow
+
+    def build_cell(cell)
+      field = super(cell)
+      field.clearButtonMode = UITextFieldViewModeNever
+      field.swizzle(:caretRectForPosition) do
+        def caretRectForPosition(position)
+          CGRectZero
+        end
+      end
+      field
+    end
+
+    def add_callbacks(field)
+      super
+      field.should_change? do |text_field|
+        false
+      end
+    end
+
+    end
+  end
+end
