+++ spec/functional/picker_row_spec.rb
@@ -58,4 +58,9 @@ describe "FormController/PickerRow" do
       end
     end
   end
+
+  it "doesn't have a clear button" do
+    tap("Picker")
+    picker_row.text_field.subviews.map(&:class).should.not.include UIButton
+  end
 end
